requirements.txt
aiobotocore==2.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.14.0
amqp==5.3.1
attrs==24.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.35.85
botocore==1.35.81
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.2
multidict==6.1.0
numpy==2.2.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
pyarrow==18.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
s3fs==2024.12.0
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
